+++ packages/common/src/weathergen/common/__init__.py
@@ -1,0 +1,2 @@
+def common_function():
+    return "This is a common function for weather generation."
