+++ pyproject.toml
@@ -25,6 +25,7 @@ dependencies = [
  "flash-attn; sys_platform == 'linux'",
  "polars~=1.25.2",
  "omegaconf~=2.3.0",
+ "weathergen-common",
 ]
 
 [project.urls]
@@ -34,13 +35,17 @@ Repository = "https://github.com/ecmwf/WeatherGenerator.git"
 Issues = "https://github.com/ecmwf/WeatherGenerator/issues"
 
 [project.scripts]
-train = "weathergen:train"
-train_continue = "weathergen:train_continue"
-evaluate = "weathergen:evaluate"
+train = "weathergen.run_train:train"
+train_continue = "weathergen.run_train:train_continue"
+evaluate = "weathergen.run_train:evaluate"
+plot = "weathergen.evaluate.plot:plot"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/weathergen"]
 
 [dependency-groups]
 # The development dependencies
@@ -120,6 +125,8 @@ url = "https://download.pytorch.org/whl/cpu"
 explicit = true
 
 [tool.uv.sources]
+weathergen-common = { workspace = true }
+weathergen-evaluate = { workspace = true }
 torch = [
   { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
   { index = "pytorch-cpu", marker = "sys_platform == 'macosx'"},
@@ -135,3 +142,10 @@ log_cli = true
 log_cli_level = "INFO"
 log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
 log_cli_date_format = "%Y-%m-%d %H:%M:%S"
+
+
+[tool.uv.workspace]
+members = [
+   "packages/evaluate",
+   "packages/common"
+]
